completed the form response flow

+++ src/components/formBuilder/FormList.js
@@ -1,28 +1,36 @@
 import React, { useState, useEffect } from 'react';
-import { Box, Button, Container, List, ListItem, ListItemText, Typography } from '@mui/material';
-import { useNavigate } from 'react-router-dom';
-import { getDocs, deleteDoc, doc } from 'firebase/firestore';
-import { formCollection } from '../../firebase'; // Firebase configuration file
+import { Box, Button, Container, List, ListItem, ListItemText, Typography, TextField } from '@mui/material';
+import { useLocation, useNavigate } from 'react-router-dom';
+import { getDocs, deleteDoc, doc, query, where } from 'firebase/firestore';
+import { formCollection } from '../../firebase';
 
 const FormList = () => {
   const [forms, setForms] = useState([]);
+  const location = useLocation();
+  const [searchTerm, setSearchTerm] = useState('');
   const navigate = useNavigate();
+  const user_id = localStorage.getItem('user_id');
+  
 
   useEffect(() => {
-    // Fetch the list of form names from the 'formSchema' collection in Firebase
-    getDocs(formCollection)
+    fetchForms();
+  }, [location]);
+
+  const fetchForms = () => {
+    getDocs(query(formCollection, where('user_id', '==',user_id)))
       .then((querySnapshot) => {
         const formList = [];
         querySnapshot.forEach((doc) => {
           const formData = doc.data();
           formList.push({ ...formData, id: doc.id });
         });
-        setForms(formList);
+        console.log(formList)
+        setForms(formList.sort((a, b) => b.modifiedAt.toDate() - a.modifiedAt.toDate()));
       })
       .catch((error) => {
         console.log('Error getting documents: ', error);
       });
-  }, []);
+  };
 
   const handleSelectForm = (form, action) => {
     if (action === 'edit') {
@@ -33,20 +41,40 @@ const FormList = () => {
   };
 
   const handleDeleteForm = async (formId) => {
-    try {
-      await deleteDoc(doc(formCollection, formId));
-      setForms(forms.filter(form => form.id !== formId));
-    } catch (error) {
-      console.log('Error deleting form: ', error);
+    const isConfirmed = window.confirm("Are you sure you want to delete this form? This action cannot be undone.");
+  
+    if (isConfirmed) {
+      try {
+        await deleteDoc(doc(formCollection, formId));
+        setForms(forms.filter(form => form.id !== formId));
+      } catch (error) {
+        console.log('Error deleting form: ', error);
+      }
+    } else {
+      // User canceled deletion
+      console.log("Form deletion canceled.");
     }
   };
+
+  const handleSearch = (e) => {
+    setSearchTerm(e.target.value);
+  };
+
+  const filteredForms = forms.filter(form => form.name.toLowerCase().includes(searchTerm.toLowerCase()));
 
   return (
     <Container>
       <Typography variant="h4" gutterBottom>Available Forms</Typography>
+      <TextField
+        fullWidth
+        label="Search Forms"
+        value={searchTerm}
+        onChange={handleSearch}
+        margin="normal"
+      />
       <List>
-        {forms.map((form) => (
-          <ListItem key={form.id} button>
+        {filteredForms.map((form) => (
+          <ListItem key={form.id}>
             <Box
               sx={{
                 display: 'flex',
@@ -60,7 +88,10 @@ const FormList = () => {
                 boxShadow: 1,
               }}
             >
-              <ListItemText primary={form.name} />
+              <ListItemText
+                primary={form.name}
+                secondary={`Modified At: ${form.modifiedAt ? new Date(form.modifiedAt.toDate()).toLocaleString() : 'N/A'}`}
+              />
               <Box>
                 <Button
                   variant="contained"
